Shots and accuracy stats for projectiles, including partial accuracy credit for splash damage.
[CL 2645096 by Steve Polge in UT-Clean branch]

+++ UnrealTournament/Source/UnrealTournament/Public/UTWeapon.h
@@ -860,8 +860,8 @@ public:
 
 	virtual int32 GetWeaponKillStats(AUTPlayerState * PS) const;
 	virtual int32 GetWeaponDeathStats(AUTPlayerState * PS) const;
-	virtual int32 GetWeaponShotsStats(AUTPlayerState * PS) const;
-	virtual int32 GetWeaponHitsStats(AUTPlayerState * PS) const;
+	virtual float GetWeaponShotsStats(AUTPlayerState * PS) const;
+	virtual float GetWeaponHitsStats(AUTPlayerState * PS) const;
 
 	// TEMP for testing 1p offsets
 	UFUNCTION(exec)
